+++ src/Contracts/CollectionTransferInterface.php
@@ -2,10 +2,11 @@
 
 namespace JuanchoSL\DataTransfer\Contracts;
 
-interface CollectionTransferInterface extends \Iterator, \Countable, \JsonSerializable
+interface CollectionTransferInterface extends \Iterator, \ArrayAccess, \Countable, \JsonSerializable
 {
 
     public function isEmpty(): bool;
+
     public function hasElements(): bool;
 
     public function append(mixed $value): int;