+++ src/DataConverters/ArrayConverter.php
@@ -4,11 +4,16 @@ declare(strict_types=1);
 
 namespace JuanchoSL\DataTransfer\DataConverters;
 
-class ArrayConverter extends AbstractConverter
+class ArrayConverter extends JsonConverter
 {
 
-    public function getData()
+    public function getData(): mixed
     {
-        return json_decode(JsonConverter::convert($this->data), true);
+        return json_decode((string) parent::getData(), true);
+    }
+
+    public function __tostring(): string
+    {
+        return serialize($this->getData());
     }
 }