+++ src/DataConverters/CsvConverter.php
@@ -4,19 +4,29 @@ declare(strict_types=1);
 
 namespace JuanchoSL\DataTransfer\DataConverters;
 
-class CsvConverter extends AbstractConverter
+class CsvConverter extends ArrayConverter
 {
 
-    public function getData()
+    protected string $separator = ',';
+
+    public function getData(): mixed
     {
-        $data = ArrayConverter::convert($this->data);
+        //$data = ArrayConverter::convert($this->data);
+        $data = parent::getData();
         if (!is_numeric(key($data))) {
             $data = [$data];
         }
         return $this->collection2csv($data);
     }
 
-    protected function array2csv($array, &$title, &$data)
+    /**
+     * Summary of array2csv
+     * @param iterable $array
+     * @param array<int|string, mixed> $title
+     * @param array<int|string, mixed> $data
+     * @return void
+     */
+    protected function array2csv(iterable $array, array &$title, array &$data): void
     {
         foreach ($array as $key => $value) {
             if (is_array($value)) {
@@ -30,7 +40,8 @@ class CsvConverter extends AbstractConverter
             }
         }
     }
-    protected function collection2csv($array)
+
+    protected function collection2csv(iterable $array): string
     {
         $results = [];
         $title = [];
@@ -48,10 +59,16 @@ class CsvConverter extends AbstractConverter
             $text = '';
             foreach ($title as $key => $value) {
                 $text .= $result[$key] ?? '';
-                $text .= ',';
+                $text .= $this->separator;
             }
             $texts[] = substr($text, 0, -1);
         }
-        return implode(',', $title) . PHP_EOL . implode(PHP_EOL, $texts);
+        return implode($this->separator, $title) . PHP_EOL . implode(PHP_EOL, $texts);
+    }
+
+    public function __tostring(): string
+    {
+        return $this->getData();
+        return chr(239) . chr(187) . chr(191) . $this->getData();
     }
 }