+++ src/DataConverters/ExcelCsvConverter.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace JuanchoSL\DataTransfer\DataConverters;
+
+class ExcelCsvConverter extends CsvConverter
+{
+    protected string $separator = ';';
+}