+++ src/DataConverters/JsonConverter.php
@@ -7,9 +7,13 @@ namespace JuanchoSL\DataTransfer\DataConverters;
 class JsonConverter extends AbstractConverter
 {
 
-    public function getData()
+    public function getData(): mixed
     {
         return json_encode($this->data);
-
+    }
+    
+    public function __tostring(): string
+    {
+        return $this->getData();
     }
 }