+++ src/DataConverters/ObjectConverter.php
@@ -4,11 +4,15 @@ declare(strict_types=1);
 
 namespace JuanchoSL\DataTransfer\DataConverters;
 
-class ObjectConverter extends AbstractConverter
+class ObjectConverter extends JsonConverter
 {
-    public function getData()
+    public function getData(): mixed
     {
-        return json_decode(JsonConverter::convert($this->data), false);
-
+        return json_decode(parent::getData(), false);
+    }
+    
+    public function __tostring(): string
+    {
+        return serialize($this->getData());
     }
 }