+++ src/DataConverters/XmlObjectConverter.php
@@ -3,26 +3,32 @@
 declare(strict_types=1);
 
 namespace JuanchoSL\DataTransfer\DataConverters;
+use JuanchoSL\DataTransfer\Contracts\DataTransferInterface;
+use JuanchoSL\Exceptions\UnprocessableEntityException;
 
 class XmlObjectConverter extends AbstractConverter
 {
 
-    public function getData()
+    public function getData(): mixed
     {
-        $key = 'root';
+        //$key = 'root';
         if ($this->data->count() == 1 && !empty($this->data->key()) && !is_numeric($this->data->key())) {
             $key = $this->data->key();
         }
-        return $this->array2XML($this->data, $key);
+        $key ??= 'root';
+        return $this->array2XML($this->data, (string) $key);
     }
 
-    protected function array2XML($data, $rootNodeName = 'root', $xml = NULL)
+    protected function array2XML(DataTransferInterface $data, string $rootNodeName = 'root', \SimpleXMLElement $xml = NULL): \SimpleXMLElement
     {
         if ($xml == null) {
             $xml = simplexml_load_string("<?xml version='1.0' encoding='utf-8'?><$rootNodeName />");
+            if ($xml === false) {
+                throw new UnprocessableEntityException("Come error creating root node {$rootNodeName}");
+            }
         }
         foreach ($data as $key => $value) {
-            if (is_iterable($value)) {
+            if ($value instanceof DataTransferInterface) {
                 if ($key == 'attributes') {
                     foreach ($value as $attr_key => $attr_value) {
                         $xml->addAttribute($attr_key, $attr_value);
@@ -42,20 +48,33 @@ class XmlObjectConverter extends AbstractConverter
                         }
                         $this->array2XML($value, $name, $node);
                     } else {
-
                         $this->array2XML($value, $key, $xml);
                     }
                 }
             } elseif (!is_numeric($key)) {
-                $value = htmlspecialchars($value);
+                //$value = htmlspecialchars($value);
                 if ($key != 'value') {
                     $node = $xml->addChild($key);
-                    $node[0] = $value;
+                    if (strpos($value, '&lt;') !== false || strpos($value, '&amp;') !== false) {
+                        $value = html_entity_decode($value);
+                    }
+                    if (strpos($value, '<') !== false || strpos($value, '&') !== false) {
+                        $node = dom_import_simplexml($node);
+                        $no = $node->ownerDocument;
+                        $node->appendChild($no->createCDATASection($value));
+                    } else {
+                        $node[0] = $value;
+                    }
                 } else {
                     $xml[0] = $value;
                 }
             }
         }
         return $xml;
+    }
+
+    public function __tostring(): string
+    {
+        return $this->getData()->asXML();
     }
 }