+++ src/Repositories/BaseCollectionable.php
@@ -6,10 +6,7 @@ namespace JuanchoSL\DataTransfer\Repositories;
 
 use JuanchoSL\DataTransfer\Contracts\CollectionTransferInterface;
 
-/**
- * @implements \Iterator<int|string, mixed>
- */
-abstract class BaseCollectionable implements CollectionTransferInterface, \ArrayAccess
+abstract class BaseCollectionable implements CollectionTransferInterface
 {
 
     /**