+++ src/Repositories/CsvDataTransfer.php
@@ -4,9 +4,17 @@ declare(strict_types=1);
 
 namespace JuanchoSL\DataTransfer\Repositories;
 
+use JuanchoSL\Exceptions\UnprocessableEntityException;
+
 class CsvDataTransfer extends ArrayDataTransfer
 {
+    protected string $separator = ',';
 
+    /**
+     * Summary of __construct
+     * @param array<int|string>|string $csv
+     * @throws \JuanchoSL\Exceptions\UnprocessableEntityException
+     */
     public function __construct(array|string $csv)
     {
         if (is_string($csv)) {
@@ -16,12 +24,20 @@ class CsvDataTransfer extends ArrayDataTransfer
                 $csv = explode(PHP_EOL, $csv);
             }
         }
+        if (!is_iterable($csv) or empty($csv)) {
+            throw new UnprocessableEntityException("No contents has been received");
+        }
+        if (($current = current($csv)) === false) {
+            throw new UnprocessableEntityException("No headers has been readed");
+        }
         $result = [];
-        $headers = str_getcsv(current($csv));
-        $csv = array_slice($csv, 1);
-        foreach ($csv as $line) {
-            $body = str_getcsv($line);
-            $result[] = array_combine($headers, $body);
+        if (count($csv) > 1) {
+            $headers = str_getcsv((string) $current, $this->separator);
+            $csv = array_slice($csv, 1);
+            foreach ($csv as $line) {
+                $body = str_getcsv((string) $line, $this->separator);
+                $result[] = array_combine($headers, $body);
+            }
         }
         parent::__construct($result);
     }